requirements/prod.txt
django
python-decouple
asgiref==3.9.1
Django==5.0
sqlparse==0.5.3